Set up Hilt and ironed out bugs

+++ app/src/main/java/com/magicgetafix/android/punkipabeerapplication/model/BeerViewModel.kt
@@ -1,5 +1,5 @@
 package com.magicgetafix.android.punkipabeerapplication.model
 
-class BeerViewModel(val id: Int, val name: String, val imageUrl: String, val strength: Int, val tagline: String = "", val description: String = "", val foodPairingNotes: List<String>) {
+class BeerViewModel(val id: Int, val name: String, val imageUrl: String, val strength: Double, val tagline: String = "", val description: String = "", val foodPairingNotes: List<String>) {
 
 }
